Added Proper structure to the project
