+++ apikeys.json
@@ -1,4 +1,4 @@
 {
-    "ad_account_id": "<ad_account_id>",
+    "ad_account_id": "act_<ad_account_id>",
     "access_token": "<access_token from Graph>"
 }
