+++ app/routes/campaign.js
@@ -1,0 +1,7 @@
+module.exports = app => {
+  const Campaign = require("../controllers/campaign.js");
+  app.get("/campaign", Campaign.getAll);
+  app.post("/campaign", Campaign.createCampaign);
+  app.delete("/campaign/:id", Campaign.deleteCampaign);
+  app.put("/campaign/:id", Campaign.updateCampaign);
+};
